+++ src/tools/goals/goalSubgoalTask.ts
@@ -16,7 +16,7 @@
 
 import { mutateGoalStatusAsync } from "../../goal/goalStore.js";
 import { addSubgoal, linkTaskToSubgoal, updateSubgoalStatus, type Subgoal } from "../../goal/goalStatus.js";
-import { createTask, type CreateTaskInput } from "../tasks/createTask.js";
+import { createTask, MODEL_SELECTION_REPO_PATH, type CreateTaskInput } from "../tasks/createTask.js";
 import { createWorktree, discardWorktree } from "../../goal/worktreeManager.js";
 import { getConfig } from "../../config.js";
 import { PatchWardenError } from "../../errors.js";
@@ -129,6 +129,7 @@ export async function createSubgoalTask(input: CreateSubgoalTaskInput): Promise<
   let taskId: string;
   try {
     const taskResult = await createTask({
+      [MODEL_SELECTION_REPO_PATH]: goalRepoPath,
       plan_id: input.plan_id,
       inline_plan: input.inline_plan,
       plan_title: input.plan_title,